more destiny updates also updated some packets, too

+++ src/eve-core/utils/utils_time.h
@@ -49,6 +49,6 @@ extern void Win32TimeToUnixTime( uint64 win32t, time_t &unix_time, uint32 &nsec 
 extern std::string Win32TimeToString(uint64 win32t);
 //extern uint32 timeNow();
 
-extern uint64 GetTimeMSeconds();
+extern double GetTimeMSeconds();
 
 #endif /* !__UTILS_TIME_H__INCL__ */
